// src/generated/src/proto/K.ts
export enum GrpcPetCategory {
  Dog = 0,
  Cat = 1,
  Bunny = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum GrpcPetStatus {
  Available = 0,
  Pending = 1,
  Sold = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface GrpcPet {
  id?: number | undefined;
  name?: string | undefined;
  category?: GrpcPetCategory | undefined;
  photoUrls: string[];
  tags?: string | undefined;
  status?: GrpcPetStatus | undefined;
}

export interface GrpcPets {
  grpcPet: GrpcPet[];
}

function createBaseGrpcPet(): GrpcPet {
  return { id: undefined, name: undefined, category: undefined, photoUrls: [], tags: undefined, status: undefined };
}

export const GrpcPet: MessageFns<GrpcPet> = {
  encode(message: GrpcPet, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      writer.uint32(8).sint64(message.id);
    }
    if (message.name !== undefined) {
      writer.uint32(18).string(message.name);
    }
    if (message.category !== undefined) {
      writer.uint32(24).int32(message.category);
    }
    for (const v of message.photoUrls) {
      writer.uint32(34).string(v!);
    }
    if (message.tags !== undefined) {
      writer.uint32(42).string(message.tags);
    }
    if (message.status !== undefined) {
      writer.uint32(48).int32(message.status);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GrpcPet {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGrpcPet();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = longToNumber(reader.sint64());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.category = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.photoUrls.push(reader.string());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.tags = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GrpcPet {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : undefined,
      name: isSet(object.name) ? globalThis.String(object.name) : undefined,
      category: isSet(object.category) ? grpcPetCategoryFromJSON(object.category) : undefined,
      photoUrls: globalThis.Array.isArray(object?.photoUrls)
        ? object.photoUrls.map((e: any) => globalThis.String(e))
        : [],
      tags: isSet(object.tags) ? globalThis.String(object.tags) : undefined,
      status: isSet(object.status) ? grpcPetStatusFromJSON(object.status) : undefined,
    };
  },

  toJSON(message: GrpcPet): unknown {
    const obj: any = {};
    if (message.id !== undefined) {
      obj.id = Math.round(message.id);
    }
    if (message.name !== undefined) {
      obj.name = message.name;
    }
    if (message.category !== undefined) {
      obj.category = grpcPetCategoryToJSON(message.category);
    }
    if (message.photoUrls?.length) {
      obj.photoUrls = message.photoUrls;
    }
    if (message.tags !== undefined) {
      obj.tags = message.tags;
    }
    if (message.status !== undefined) {
      obj.status = grpcPetStatusToJSON(message.status);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GrpcPet>, I>>(base?: I): GrpcPet {
    return GrpcPet.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GrpcPet>, I>>(object: I): GrpcPet {
    const message = createBaseGrpcPet();
    message.id = object.id ?? undefined;
    message.name = object.name ?? undefined;
    message.category = object.category ?? undefined;
    message.photoUrls = object.photoUrls?.map((e) => e) || [];
    message.tags = object.tags ?? undefined;
    message.status = object.status ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}